requirements.txt
Pillow
customtkinter>=5.2.0